Refactored more items into folders so that it's easier to tell what each component is used/used for.
