+++ src/components/App.js
@@ -10,18 +10,18 @@ import { AuthProvider } from "../contexts/AuthContext";
 import Signin from "./Authentication/Signin";
 import Signout from "./Authentication/Signout";
 import SignupPage from "./Authentication/SignupPage";
-import PrivateRoute from "./Helpers/PrivateRoute";
-import Header from "./Header/Header";
+import PrivateRoute from "./Utility/Helpers/PrivateRoute";
+import Header from "./Landing/Header/Header";
 import "./App.css";
-import HomePage from "./HomePage/HomePage";
-import Progress from "./Progress/Progress";
-import ChallengerWelcome from "./ChallengerWelcome/ChallengerWelcome";
-import ElectionReminder from "./ElectionReminders/ElectionReminder";
-import Footer from "./Footer/Footer";
-import PlayerWelcome from "./PlayerWelcome/PlayerWelcome";
+import HomePage from "./Landing/HomePage/HomePage";
+import Progress from "./Challenge/Progress/Progress";
+import ChallengerWelcome from "./Challenge/ChallengerWelcome/ChallengerWelcome";
+import ElectionReminder from "./Challenge/ElectionReminders/ElectionReminder";
+import Footer from "./Landing/Footer/Footer";
+import PlayerWelcome from "./Landing/PlayerWelcome/PlayerWelcome";
 import Actions from "./Actions/Actions";
-import TOS from "./Legal/TOS";
-import PrivacyPolicy from "./Legal/PrivacyPolicy";
+import TOS from "./Landing/Legal/TOS";
+import PrivacyPolicy from "./Landing/Legal/PrivacyPolicy";
 //Voter Registration Pages
 import { Eligibility } from "./VoterRegistration/pages/Eligibility/Eligibility.component";
 import { YourName } from "./VoterRegistration/pages/YourName/YourName.component";
