+++ src/components/Challenge/ChallengePage/ChallengeIcon.js
