+++ src/components/Challenge/Invite/Instagram/Instagram.scss
