+++ src/components/Challenge/Progress/Progress.scss
@@ -1,0 +1,122 @@
+@import "src/styles/main";
+
+.progress-page {
+  .section-1 {
+    .days-blob-container {
+      height: 0px;
+      bottom: 110px;
+      float: right;
+      position: relative;
+      .days-label {
+        @include horizontal-center;
+        flex-direction: column;
+        position: absolute;
+        left: 45%;
+        top: 40px;
+        p {
+          font-size: 122px;
+          font-family: $bebas-neue;
+          line-height: 110px;
+          margin: 0;
+        }
+      }
+      .blob {
+        width: 180px;
+        height: 220px;
+      }
+    }
+    @include xs-screen {
+      .days-blob-container {
+        .days-label {
+          left: 45%;
+          top: 40px;
+          p {
+            font-size: 132px;
+            line-height: 120px;
+          }
+        }
+        .blob {
+          width: 200px;
+          height: 220px;
+        }
+      }
+    }
+  }
+
+  .section-2 {
+    span {
+      font-size: 48px;
+    }
+  }
+
+  .section-3 {
+    display: -ms-grid;
+    display: grid;
+    -ms-grid-columns: auto auto;
+    grid-template-columns: auto auto;
+    overflow: hidden;
+    .badges {
+      @include center;
+      margin-bottom: 32px;
+      .blob {
+        @include center;
+        position: relative;
+        z-index: 0;
+        width: 0px;
+        img {
+          width: 182px;
+          height: 183px;
+        }
+      }
+      .disable {
+        opacity: 0.2;
+      }
+      .blob-content {
+        @include center;
+        flex-direction: column;
+        position: relative;
+        z-index: 0.1;
+        width: 0;
+        .icon {
+          width: 100px;
+          height: 100px;
+          margin-bottom: 8px;
+        }
+        .index {
+          color: $white;
+          font-size: 72px;
+          font-family: $bebas-neue;
+          line-height: 120px;
+          margin: 0;
+        }
+        h3 {
+          width: 120px;
+          text-align: center;
+        }
+      }
+    }
+
+    .badges {
+      .blob-img {
+        position: relative;
+      }
+
+      .blob-img-1 {
+        right: 16px;
+      }
+
+      .blob-img-3 {
+        right: 16px;
+      }
+
+      .blob-img-7 {
+        bottom: 10px;
+        right: 10px;
+      }
+
+      .blob-img-8 {
+        left: 20px;
+      }
+    }
+  }
+}
