--- src/components/Invite/Components/Preview.scss
@@ -1,85 +1,0 @@
-@import "../../../styles/Fonts.scss";
-
-.preview-wrapper {
-  width: 0;
-  height: 95%;
-  max-width: 500px;
-  position: fixed;
-  bottom: -100vh;
-  transition: 400ms;
-  background-color: white;
-  color: black;
-  overflow: auto;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-family: $lato;
-  border-top: 2px solid black;
-
-  .preview-header {
-    width: 100%;
-    height: 60px;
-    display: flex;
-    margin-top: 50px;
-
-    a.nav-link,
-    a.nav-link:hover {
-      color: black;
-      fill: black;
-      margin: 25px;
-    }
-
-    #close-icon {
-      position: relative;
-      right: 20px;
-      bottom: 20px;
-      svg {
-        width: 30px;
-        height: 30px;
-      }
-    }
-
-    .invite-title {
-      position: relative;
-      right: 40px;
-      bottom: 5px;
-    }
-  }
-
-  .invite-title {
-    font-family: $bebas-neue;
-    margin: auto;
-    font-size: 36px;
-  }
-
-  .view {
-    width: 100%;
-    display: -ms-grid;
-    display: grid;
-    justify-items: center;
-
-    .preview-bordor {
-      grid-column: 1;
-      grid-row: 1;
-      width: 335px;
-    }
-
-    .playerwelcome-page {
-      overflow: auto;
-      grid-column: 1;
-      grid-row: 1;
-      width: 255px;
-      height: 500px;
-      margin-top: 93px;
-      margin-left: 7px;
-      border-radius: 15px 15px 40px 42px;
-    }
-  }
-}
-
-.preview-wrapper.active {
-  bottom: 0;
-  transition: 400ms;
-  width: 100%;
-}
