+++ src/components/Landing/Footer/Footer.scss
