+++ src/components/Landing/Legal/Legal.scss
@@ -1,0 +1,17 @@
+@import "src/styles/Fonts";
+
+.Legal {
+  padding: 30px;
+
+  h1 {
+    margin-bottom: 10px;
+  }
+
+  h2 {
+    margin-bottom: 5px;
+  }
+
+  p {
+    margin-bottom: 20px;
+  }
+}
