--- src/components/Legal/Legal.scss
@@ -1,17 +1,0 @@
-@import "../../styles/Fonts.scss";
-
-.Legal {
-  padding: 30px;
-
-  h1 {
-    margin-bottom: 10px;
-  }
-
-  h2 {
-    margin-bottom: 5px;
-  }
-
-  p {
-    margin-bottom: 20px;
-  }
-}
