--- src/components/PlayerWelcome/PlayerWelcome.scss
@@ -1,95 +1,0 @@
-@import "./../../styles/Components.scss";
-
-.player-welcome {
-  background-color: $white;
-  color: $black;
-  display: flex;
-  flex-direction: column;
-
-  .getStarted-button {
-    @include button;
-    outline: 4px solid white;
-    margin-top: 25px;
-  }
-
-  .top {
-    display: flex;
-    flex-direction: column;
-    padding: 30px;
-    background-color: $black;
-    color: white;
-    .top-heading {
-      @include heading-text(white, 50px);
-      line-height: 57.6px;
-      letter-spacing: 0.03em;
-      margin: 20px 0px;
-      .underline {
-        @include text-underline;
-      }
-    }
-  }
-
-  .signin-link {
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-    text-decoration: underline;
-    display: inline;
-    margin: 0;
-    padding: 0px 0px 10px 0px;
-    font-family: $lato;
-    font-size: 14px;
-    line-height: 22px;
-    letter-spacing: 0.02em;
-    font-weight: 900;
-  }
-
-  .blue {
-    color: $blue;
-  }
-
-  .black {
-    color: $black;
-  }
-
-  .main-content {
-    padding: 40px 30px;
-
-    h3 {
-      align-self: flex-start;
-    }
-  }
-
-  .heading {
-    @include heading-text($black, 36px);
-    margin-bottom: 40px;
-    line-height: 43.2px;
-    letter-spacing: 0.03em;
-    text-align: center;
-  }
-
-  .text {
-    font-family: $lato;
-    font-size: 18px;
-    line-height: 22px;
-    letter-spacing: 0.02em;
-    margin-bottom: 32px;
-  }
-
-  .small-text {
-    font-family: $lato;
-    font-size: 15px;
-    letter-spacing: 0.02em;
-    margin-top: 15px;
-  }
-
-  .curve {
-    width: 100%;
-  }
-
-  .image {
-    padding: 20px 0px;
-    padding-bottom: 60px;
-    text-align: center;
-  }
-}
